remove hardcoding and unused code
